large code update
 - initial migration code to symphonia & cpal backend
   - symphonia boilerplate
   - motivation to remove dependency on rodio
 - add support for getting audio duration via symphonia
 - display audio system being used (mostly for debug purposes right now)

+++ lib/dizi_lib/src/response/server.rs
@@ -25,7 +25,7 @@ pub enum ServerBroadcastEvent {
     PlayerShuffle { on: bool },
     PlayerNext { on: bool },
 
-    PlayerVolumeUpdate { volume: usize },
+    PlayerVolumeUpdate { volume: f32 },
     PlayerProgressUpdate { elapsed: time::Duration },
 
     // playlist
